requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.4
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.1.0
django-reversion==3.0.2
djangorestframework==3.9.0
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.1.
odfpy==1.4.0
openpyxl==2.5.12
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
xadmin==2.0.1
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
